refactor: update component imports to be direct and not barrel
