+++ src/app/(locale)/(public)/cart/page.tsx
@@ -1,6 +1,7 @@
 "use client";
 
-import { QtySelector, Title } from "@/components";
+import { QtySelector } from "@/components/product/qty-selector/qty-selector";
+import { Title } from "@/components/ui/title/title";
 import Image from "next/image";
 import Link from "next/link";
 import { redirect } from "next/navigation";
