+++ src/app/(locale)/(public)/category/not-found.tsx
@@ -1,4 +1,4 @@
-import { NotFound } from "@/components";
+import { NotFound } from "@/components/ui/not-found/not-found";
 
 export default function NotFoundPage() {
   return <NotFound />;
