+++ src/app/(locale)/(user)/checkout/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import { Title } from "@/components";
+import { Title } from "@/components/ui/title/title";
 import { useToast } from "@/components/ui/toast/toast-provider";
-import { useCartStore } from "@/store";
+import { useCartStore } from "@/store/cart/cart-store";
 import { useOrderStore } from "@/store/order/order-store";
 import Image from "next/image";
 import Link from "next/link";
