+++ src/app/(locale)/(user)/checkout/shipment/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { Title } from "@/components";
+import { Title } from "@/components/ui/title/title";
 import { useState } from "react";
 import { useOrderStore } from "@/store/order/order-store";
 import { redirect } from "next/navigation";
