--- src/components/index.ts
@@ -1,31 +1,0 @@
-// Navigation & Layout
-export * from "./ui/navbar/navbar";
-export * from "./ui/sidebar/sidebar";
-export * from "./ui/footer/footer";
-
-// Content & Display
-export * from "./ui/title/title";
-export * from "./ui/not-found/not-found";
-
-// Products
-export * from "./products/product-grid/product-grid";
-export * from "./products/product-grid/product-item";
-export * from "./product/size-selector/size-selector";
-export * from "./product/qty-selector/qty-selector";
-export * from "./product/slideshow/product-slideshow";
-export * from "./product/slideshow/product-slideshow-mobile";
-
-// UI Components
-export * from "./ui/button/button";
-export * from "./ui/input/input";
-export * from "./ui/card/card";
-export * from "./ui/badge/badge";
-export * from "./ui/avatar/avatar";
-export * from "./ui/loading/loading";
-export * from "./ui/toggle-button/toggle-button";
-export * from "./ui/toast/toast-provider";
-export * from "./ui/error-boundary/error-boundary";
-
-// Forms
-export * from "./ui/form/login-form";
-export * from "./ui/form/signup-form";
