--- src/store/index.ts
@@ -1,3 +1,0 @@
-export * from "./ui/ui-store";
-export * from "./cart/cart-store";
-export * from "./order/order-store";
